requirements.txt
torch
torchvision
scikit-learn
opencv-python
matplotlib
numpy
seaborn
tqdm
pillow
streamlit
